Update
> Deleted the comment @Bug from the files.

>UUID change the constructor with int32 to uint64.

>Added the function destroy() in the next classes:
  -StaticMesh
  -SceneNode
  -Component

>Added the function clearGraph() in the class SceneGraph.

>Changed the parameter of the function getActorByName() in the class SceneNode to receive a StringView.

>Changed the parameter of the function getActorByID*( in the class SceneNode to receive a const int32&.

>Changed the name of the function readFromFile() in the class ResourceManager to resourceFromFile().

>Added the function readFromFile.

>Implemented the function encodeGiProject() in the class Encoder.

>Added the funcion encodeGiScene() in the class Encoder.

>Added the function decodeFile() in the class Decoder.

>Added the namespace/Enum CAMERA_TYPE in the file giCamera.h.

>Added the variable m_cameraType in the class Camera.

>Added the next fuctions to the class Editor:
  -OpenFileDialog
  -SaveFileDialog

>Created the file giBaseImplementations.h and it .cpp.

>Created the class BaseFileDialogs.

>Created the file giSpecificImplementations.h and it .cpp.

>Created the file giWindowsImplementations.h and it .cpp.

>Created the class WindowsFileDialogs.

>Added ofstream to the giStdHeaders.h.

+++ giCore/include/giUUID.h
@@ -3,8 +3,7 @@
  * @author  Jesus Alberto Del Moral Cupil
  * @e       idv18c.jmoral@uartesdigitales.edu.mx
  * @date    09/08/2022
- * @brief   Universal Unique Identifiers.
- * @bug     No known Bugs.
+ * @brief   An implementation for the Universal Unique IDentifiers.
  */
  
 /**
@@ -22,7 +21,7 @@ namespace giEngineSDK {
     UUID();
 
     //Constructor for own purposes.
-    UUID(int32 inData);
+    UUID(uint64 inData);
 
     //Copy constructor.
     UUID(const UUID& inUUID) = default;
